error handling

+++ src/Pages/Dashboard/ManageOrderTable/ManageOrderTable.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Container } from 'react-bootstrap';
+import swal from 'sweetalert';
 
 
 const ManageOrderTable = (props) => {
@@ -7,22 +8,34 @@ const ManageOrderTable = (props) => {
 
     const handleDelete=id=>{
 
-      fetch(`http://localhost:5000/orders/${id}`,{
-        method:'DELETE',
-        headers:{
-          'content-type':'application/json'
-        }
+      swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this imaginary file!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
       })
-      .then(res=>res.json()
-      .then(result=>{
-
-      }))
+      .then((willDelete) => {
+        if (willDelete) {
+          fetch(`http://localhost:5000/orders/${id}`, {
+        method: "DELETE",
+        headers: { "Content-type": "application/json" },
+        
+      })
+        .then((res) => res.json())
+        .then((result) => {
+             
+        });
+        } else {
+          swal("Your imaginary file is safe!");
+        }
+      });
 
     };
 
     const handleUpdate=id=>{
        
-      fetch(`http://localhost:5000/orders/${id}`,{
+      /* fetch(`http://localhost:5000/orders/${id}`,{
         method:'PUT',
         headers:{
           'content-type':'application/json'
@@ -31,7 +44,29 @@ const ManageOrderTable = (props) => {
       .then(res=>res.json()
       .then(result=>{
 
-      }))
+      })) */
+      swal({
+        title: "Are you sure?",
+        text: "Once Approved, you will not be able to recover this imaginary file!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      })
+      .then((willDelete) => {
+        if (willDelete) {
+          fetch(`http://localhost:5000/orders/${id}`, {
+        method: "PUT",
+        headers: { "Content-type": "application/json" },
+        
+      })
+        .then((res) => res.json())
+        .then((result) => {
+             
+        });
+        } else {
+          swal("Your imaginary file is safe!");
+        }
+      });
